Add files via upload
Based on the feedback from the presentation, I added some coding to the time series analysis to list the total number of travelers with Visitor Visas for each year (excluding 2024, as the data is incomplete) and calculated the proportion comparing 2023 with 2019. I have also added relevant explanations to the report draft and listed the datasets used in the draft report.